Extract drawdown analytics router
