+++ src/app/routers/drawdown.py
@@ -1,0 +1,68 @@
+from fastapi import APIRouter, Request
+
+from app.api_errors import STANDARD_ERROR_RESPONSES
+from app.contracts.drawdown import (
+    DrawdownAnalyticsRequest,
+    DrawdownInputMode,
+    DrawdownResponse,
+)
+from app.integrations.lotus_performance_client import LotusPerformanceClient
+from app.services.drawdown_engine import calculate_drawdown
+from app.services.drawdown_mode_adapter import calculate_drawdown_stateful
+from app.services.endpoint_observation import observed_endpoint
+
+router = APIRouter(tags=["risk-analytics"])
+
+
+@router.post(
+    "/analytics/risk/drawdown",
+    response_model=DrawdownResponse,
+    responses=STANDARD_ERROR_RESPONSES,
+    summary="Calculate realized drawdown analytics",
+    description=(
+        "Calculates realized drawdown analytics for stateless or stateful return histories, including "
+        "max drawdown, episode diagnostics, time-under-water, ulcer index, drawdown-at-risk, and "
+        "benchmark-relative drawdown timing. The response echoes applied analysis and benchmark policy "
+        "settings so downstream consumers can interpret results without reconstructing the request."
+    ),
+)
+async def analytics_risk_drawdown(
+    request_payload: DrawdownAnalyticsRequest,
+    request: Request,
+) -> DrawdownResponse:
+    input_mode = request_payload.input_mode.value
+    if request_payload.input_mode == DrawdownInputMode.STATELESS:
+        stateless_input = request_payload.stateless_input
+        assert stateless_input is not None
+        return await observed_endpoint(
+            endpoint="drawdown",
+            input_mode=input_mode,
+            operation=lambda: calculate_drawdown(
+                stateless_input,
+                input_mode=DrawdownInputMode.STATELESS,
+                analysis_options=request_payload.analysis_options,
+                include_benchmark=request_payload.benchmark_policy.include_benchmark,
+                missing_benchmark_policy=request_payload.benchmark_policy.missing_benchmark_policy,
+            ),
+        )
+
+    if request_payload.input_mode == DrawdownInputMode.STATEFUL:
+        stateful_input = request_payload.stateful_input
+        assert stateful_input is not None
+        performance_client = getattr(request.app.state, "lotus_performance_client", None)
+        if performance_client is None:
+            performance_client = LotusPerformanceClient()
+        return await observed_endpoint(
+            endpoint="drawdown",
+            input_mode=input_mode,
+            operation=lambda: calculate_drawdown_stateful(
+                stateful_input,
+                analysis_options=request_payload.analysis_options,
+                performance_client=performance_client,
+                correlation_id=request.headers.get("X-Correlation-Id"),
+            ),
+        )
+
+    raise ValueError(
+        f"Unsupported input_mode={request_payload.input_mode.value} for /analytics/risk/drawdown"
+    )
